requirements.txt
Flask==3.0.3
python-telegram-bot==21.4
requests==2.32.3
python-dotenv==1.0.1
